Add rate limiter and refactor game card constants
Introduce a thread-safe rate limiting service used by API endpoints and refactor UI constants.

- Add MemoryCacheRateLimitService (IRateLimitService) with atomic counters to prevent race conditions and register it in DI.
- Update GetPlayStationProfile and GetXboxProfile to use the new rate limit service.
- Remove deprecated .github workflow (api.yml) and adjust azure-static-web-app.yml: remove markdown path filtering and add a .NET API build step.
- Extract platform/status/localization constants from GameCard and GameCardGroup into src/components/VideoGames/gameCardConstants.js and update imports.
- Fix typo in videogames/xbox index (Theater of Idols).

These changes centralize rate limiting, avoid memory-cache races, simplify component constants, and ensure the API is built during the static web app workflow.
