--- .github/workflows/api.yml
@@ -1,52 +1,0 @@
-# =============================================================================
-# DEPRECATED: This workflow is no longer used for deployment.
-# The API is now deployed as part of Azure Static Web Apps managed functions.
-# See: azure-static-web-app.yml
-# =============================================================================
-# This workflow is kept for backward compatibility and will validate API builds
-# when changes are made to the api/ directory.
-# =============================================================================
-
-name: Validate API Build (Deprecated)
-
-on:
-  pull_request:
-    types: [opened, synchronize, reopened]
-    branches:
-      - main
-    paths:
-      - 'api/**'
-  workflow_dispatch:
-
-permissions:
-  contents: read
-
-jobs:
-  validate:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check Out Repo
-        uses: actions/checkout@v4
-
-      - name: Setup .NET
-        uses: actions/setup-dotnet@v4
-        with:
-          dotnet-version: '9.0.x'
-
-      - name: Restore Dependencies
-        working-directory: ./api
-        run: dotnet restore
-
-      - name: Build API
-        working-directory: ./api
-        run: dotnet build --configuration Release --no-restore
-
-      - name: Validate Build Output
-        working-directory: ./api
-        run: |
-          if [ -f "bin/Release/net9.0/api.dll" ]; then
-            echo "✅ API build successful"
-          else
-            echo "❌ API build failed - output not found"
-            exit 1
-          fi
