+++ videogames/xbox/index.mdx
@@ -52,7 +52,7 @@ Progress, achievements, and saves sync seamlessly across all devices. Start a ga
 
 <GameCard title="Reanimal" platform="xbox" status="completed" imageUrl="/img/videogames/xbox/reanimal.jpg" recommendation="A charming puzzle adventure." url="https://www.xbox.com/games/store/reanimal-digital-deluxe-edition/9NB3XVDJHSM8/0010" />
 
-<GameCard title="Crisol: Theather of Idols" platform="xbox" status="playing" imageUrl="/img/videogames/xbox/crisol.jpg" recommendation="A unique puzzle adventure." url="https://www.xbox.com/games/store/crisol-theater-of-idols/9N6QTNWJM2NW/0010" />
+<GameCard title="Crisol: Theater of Idols" platform="xbox" status="playing" imageUrl="/img/videogames/xbox/crisol.jpg" recommendation="A unique puzzle adventure." url="https://www.xbox.com/games/store/crisol-theater-of-idols/9N6QTNWJM2NW/0010" />
 
 <GameCardGroup title="Indiana Jones and the Great Circle" platform="xbox" recommendation="Classic adventure with Indiana Jones - base game and DLC.">
   <GameCard title="Indiana Jones and the Great Circle" platform="xbox" status="completed" imageUrl="/img/videogames/xbox/indiana-jones.jpg" url="https://www.xbox.com/games/store/indiana-jones-and-the-great-circle/9N8FQ28Z6QX3/0010" />
